requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.3.2
numpy==1.24.3
Pillow==10.1.0
bcrypt==4.1.1
email-validator==2.1.0
python-socketio==5.10.0
eventlet==0.33.3
fpdf==1.7.2
psycopg2-binary==2.9.9

